requirements.txt
python-telegram-bot[job-queue]>=20.0
aiosqlite>=0.19.0
opencc-python-reimplemented>=0.1.7
